+++ src/IMS/core/IMS.BusinessModel/ViewModel/ProductAddModel.cs
@@ -16,7 +16,7 @@ namespace IMS.BusinessModel.ViewModel
         public long CategoryId { get; set; }
         public long BrandId { get; set; }
         public string Description { get; set; }
-        public int Rank { get; set; }
+        public long Rank { get; set; }
         public Status Status { get; set; }
     }
 }
